pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["LocalInterpreter"]  # プロジェクトディレクトリに応じて変更

[project]
name = "crab-interpreter"
version = "0.0.1"
authors = [  { name="route250" },]
description = "local interpreter service"
license = {text = "MIT License"}
requires-python = ">=3.10.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    ]
dependencies = [
    "python-dotenv", "Quart", "bs4", "google-api-python-client", "pytrends", "openai", "tiktoken"
    ]